+++ src/sinks/prelude.rs
@@ -3,9 +3,14 @@
 
 pub use crate::{
     codecs::{Encoder, EncodingConfig, Transformer},
+    components::validation::{
+        ExternalResource, HttpResourceConfig, ResourceDirection, ValidatableComponent,
+        ValidationConfiguration,
+    },
     config::{DataType, GenerateConfig, SinkConfig, SinkContext},
     event::{Event, LogEvent},
     internal_events::{SinkRequestBuildError, TemplateRenderingError},
+    register_validatable_component,
     sinks::{
         util::{
             builder::SinkBuilderExt,
@@ -35,7 +40,7 @@ pub use vector_common::{
 pub use vector_config::configurable_component;
 
 pub use vector_core::{
-    config::{AcknowledgementsConfig, Input},
+    config::{telemetry, AcknowledgementsConfig, Input},
     event::Value,
     partition::Partitioner,
     schema::Requirement,
