+++ website/cue/reference/components/sinks/base/http.cue
@@ -305,9 +305,10 @@ base: components: sinks: http: configuration: {
 		}
 	}
 	headers: {
-		deprecated:  true
-		description: "A list of custom headers to add to each request."
-		required:    false
+		deprecated:         true
+		deprecated_message: "This option has been deprecated, use `request.headers` instead."
+		description:        "A list of custom headers to add to each request."
+		required:           false
 		type: object: options: "*": {
 			description: "An HTTP request header and it's value."
 			required:    true
